+++ sdk/include/TosResponse.h
@@ -57,16 +57,16 @@ public:
         Id2_ = Id2;
     }
     uint64_t getHashCrc64Result() const {
-        return hashCrc64Result;
+        return hashCrc64Result_;
     }
-    void setHashCrc64Result(uint64_t hashcrc64result) {
-        hashCrc64Result = hashcrc64result;
+    void setHashCrc64Result(uint64_t hashCrc64Result) {
+        hashCrc64Result_ = hashCrc64Result;
     }
     int getCurlErrCode() const {
-        return curlErrCode;
+        return curlErrCode_;
     }
     void setCurlErrCode(int curlErrCode) {
-        TosResponse::curlErrCode = curlErrCode;
+        curlErrCode_ = curlErrCode;
     }
 
 private:
@@ -76,7 +76,7 @@ private:
     std::map<std::string, std::string> headers_;
     std::shared_ptr<std::iostream> content_;
     std::string Id2_;
-    uint64_t hashCrc64Result = 0;
-    int curlErrCode = 0;
+    uint64_t hashCrc64Result_ = 0;
+    int curlErrCode_ = 0;
 };
 }  // namespace VolcengineTos